backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
playwright==1.40.0
python-dotenv==1.0.0
redis==5.0.1
rq==1.15.1
fake-useragent==1.4.0
openai==1.6.1
beautifulsoup4==4.12.2
aiohttp==3.9.0
ddddocr==1.4.7
cryptography==41.0.7
pyyaml==6.0.1
pydantic==2.5.0
websockets==12.0
asyncio-mqtt==0.16.1